+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- Add `hbase.rest.hostname`, `hbase.rest.port`, and `hbase.rest.info.port` properties to the restserver `hbase-site.xml` ([#708]).
+
+[#708]: https://github.com/stackabletech/hbase-operator/pull/708
+
 ## [25.11.0] - 2025-11-07
 
 ## [25.11.0-rc1] - 2025-11-06